mail-ai-backend/services/event_processor/requirement.txt
google-cloud-pubsub
google-auth
google-api-python-client
motor
google-generativeai
python-dotenv